moved tests to relational directive package
